Tuckshop multichoice function
This function allows me to copy and paste into any function that needs multiple choice. Then by using command r (ctrl r on windows), I could just replace anything with ABC with the food that was multi-choice. It was great and saved time as I didn't have to re write the whole thing and didn't have to replace any of the code through long process